backend/foodgram/users/views.py
from django.contrib.auth.models import User
from djoser.serializers import UserSerializer
from rest_framework import viewsets


class UserViewSet(viewsets.ModelViewSet):
    serializer_class = UserSerializer
    queryset = User.objects.all()
