backend/requirements.txt
djoser==2.1.0
djangorestframework==3.14.0
Django==3.2.18
